+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # dev
 
 * exclude group-css-media-queries from slim build
+* dont run bower when rerunning the generator and bower is not included
 * added also twig as static page generator
 * added script to run non-interactive-generator for builds